+++ src/main/java/com/cardano/explorer/model/response/tx/TxMintingResponse.java
@@ -3,7 +3,7 @@ package com.cardano.explorer.model.response.tx;
 import com.cardano.explorer.model.response.token.TokenMetadataResponse;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -14,7 +14,7 @@ import lombok.Setter;
 @JsonInclude(Include.NON_NULL)
 public class TxMintingResponse {
   private String assetName;
-  private BigDecimal assetQuantity;
+  private BigInteger assetQuantity;
   private String assetId;
   private String policy;
   private TokenMetadataResponse metadata;
